use meshes from gltf file individually
