+++ src/app/visualization/three-fiber-components/ResourceNode.tsx
@@ -1,16 +1,12 @@
-import React, { useRef } from 'react';
-import type { Mesh } from 'three';
+import React from 'react';
 import { MeshProps } from 'react-three-fiber';
 import { positionArgsFromCoords, ResourceNode } from 'src/game';
+import { useGLTFData } from './Asset';
 
-const ResourceNodeComponent: React.FC<MeshProps & ResourceNode> = (props) => {
-  const mesh = useRef<Mesh>();
-  return (
-    <mesh {...props} {...positionArgsFromCoords(props.coords)} ref={mesh} scale={[1, 1, 1]}>
-      <boxBufferGeometry args={[1, 1, 1]} />
-      <meshStandardMaterial color={'blue'} />
-    </mesh>
-  );
+const ResourceNodeComponent: React.FC<MeshProps & ResourceNode> = ({ coords, ...meshProps }) => {
+  const adjustedProps = { ...meshProps, ...positionArgsFromCoords(coords) };
+  const { nodes, materials } = useGLTFData();
+  return <mesh material={materials['Vegetation']} geometry={nodes['Bushes'].geometry} {...adjustedProps} />;
 };
 
 export default ResourceNodeComponent;
